Add file exists check
Glob.sync (used by sass-resource-loader) does not fail if it is unable to find a match using, instead returning an empty array. This can cause confusion if there is a typo, etc in configuration.

This is just an idea for a solution however feel free to update however you like.